Move cloversim ros tool to tool (3)
Implement command for rebuilding cloversim layer, there is also fast mode which doesn't delete previous built layer, that allows cmake to use configuration cache

+++ src/prepare.go
@@ -42,12 +42,14 @@ func Prepare() error {
 	return nil
 }
 
-func BuildCloversimLayer() error {
+func BuildCloversimLayer(buildFast bool) error {
 	HostLogger.Info("Building cloversim layer, this may take a while...")
 
 	layerPath := path.Join(LocateSetup(), "base_fs", "cloversim")
 	buildContainerPath := path.Join(LocateSetup(), "containers", "__build_cloversim")
-	os.RemoveAll(layerPath)
+	if !buildFast {
+		os.RemoveAll(layerPath)
+	}
 	os.MkdirAll(layerPath, os.ModePerm)
 
 	overlay := &Overlay{
